move SearchBox and deps to search-ui package
