+++ client/search-ui/src/index.ts
@@ -1,1 +1,3 @@
 export * from './components'
+export * from './input/toggles'
+export * from './input/SearchBox'
