+++ client/search-ui/src/input/HighlightedSearchContextSpec.tsx
