+++ client/search-ui/src/input/LazyMonacoQueryInput.module.scss
