+++ client/search-ui/src/input/LazyMonacoQueryInput.tsx
