+++ client/search-ui/src/input/MonacoQueryInput.module.scss
