+++ client/search-ui/src/input/MonacoQueryInput.tsx
