+++ client/search-ui/src/input/SearchBox.module.scss
