+++ client/search-ui/src/input/SearchContextCtaPrompt.module.scss
