+++ client/search-ui/src/input/SearchContextCtaPrompt.story.tsx
@@ -1,10 +1,9 @@
 import { storiesOf } from '@storybook/react'
 import React from 'react'
 
+import { BrandedStory } from '@sourcegraph/branded/src/components/BrandedStory'
+import { AuthenticatedUser } from '@sourcegraph/shared/src/auth'
 import { NOOP_TELEMETRY_SERVICE } from '@sourcegraph/shared/src/telemetry/telemetryService'
-
-import { AuthenticatedUser } from '../../auth'
-import { WebStory } from '../../components/WebStory'
 
 import { SearchContextCtaPrompt } from './SearchContextCtaPrompt'
 
@@ -41,7 +40,7 @@ const authUser: AuthenticatedUser = {
 add(
     'not authenticated',
     () => (
-        <WebStory>
+        <BrandedStory>
             {() => (
                 <SearchContextCtaPrompt
                     telemetryService={NOOP_TELEMETRY_SERVICE}
@@ -50,7 +49,7 @@ add(
                     onDismiss={() => {}}
                 />
             )}
-        </WebStory>
+        </BrandedStory>
     ),
     {}
 )
@@ -58,7 +57,7 @@ add(
 add(
     'authenticated without added external services',
     () => (
-        <WebStory>
+        <BrandedStory>
             {() => (
                 <SearchContextCtaPrompt
                     telemetryService={NOOP_TELEMETRY_SERVICE}
@@ -67,7 +66,7 @@ add(
                     onDismiss={() => {}}
                 />
             )}
-        </WebStory>
+        </BrandedStory>
     ),
     {}
 )
@@ -75,7 +74,7 @@ add(
 add(
     'authenticated with added external services',
     () => (
-        <WebStory>
+        <BrandedStory>
             {() => (
                 <SearchContextCtaPrompt
                     telemetryService={NOOP_TELEMETRY_SERVICE}
@@ -84,7 +83,7 @@ add(
                     onDismiss={() => {}}
                 />
             )}
-        </WebStory>
+        </BrandedStory>
     ),
     {}
 )
@@ -92,7 +91,7 @@ add(
 add(
     'authenticated with private code',
     () => (
-        <WebStory>
+        <BrandedStory>
             {() => (
                 <SearchContextCtaPrompt
                     telemetryService={NOOP_TELEMETRY_SERVICE}
@@ -101,7 +100,7 @@ add(
                     onDismiss={() => {}}
                 />
             )}
-        </WebStory>
+        </BrandedStory>
     ),
     {}
 )
