+++ client/search-ui/src/input/SearchContextCtaPrompt.tsx
