+++ client/search-ui/src/input/SearchContextDropdown.module.scss
