+++ client/search-ui/src/input/SearchHelpDropdownButton.tsx
