+++ client/search-ui/src/input/toggles/CopyQueryButton.tsx
