+++ client/search-ui/src/input/toggles/QueryInputToggle.tsx
