+++ client/search-ui/src/input/toggles/Toggles.module.scss
