+++ client/web/src/enterprise/insights/components/form/monaco-field/MonacoField.tsx
@@ -4,11 +4,11 @@ import * as Monaco from 'monaco-editor'
 import React, { createContext, forwardRef, InputHTMLAttributes, useContext, useImperativeHandle, useMemo } from 'react'
 
 import { QueryChangeSource } from '@sourcegraph/search'
+import { LazyMonacoQueryInput } from '@sourcegraph/search-ui/src/input/LazyMonacoQueryInput'
+import { DEFAULT_MONACO_OPTIONS } from '@sourcegraph/search-ui/src/input/MonacoQueryInput'
 import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 import { ForwardReferenceComponent } from '@sourcegraph/wildcard'
 
-import { LazyMonacoQueryInput } from '../../../../../search/input/LazyMonacoQueryInput'
-import { DEFAULT_MONACO_OPTIONS } from '../../../../../search/input/MonacoQueryInput'
 import { ThemePreference } from '../../../../../stores/themeState'
 import { useTheme } from '../../../../../theme'
